Rename module #136
