+++ scigateway_auth/src/icat_client.py
